refactor(app): extract site config and page-context helpers to config.py
Move SITE_CONFIG, NAV_LINKS, SITE_LINKS, build_absolute_url,
build_social_image_url, and build_page_context out of app.py into a
dedicated config module, keeping app.py focused on Flask wiring and
routes. SITE_CONFIG stays importable from app.py via re-export so
freeze.py's `from app import SITE_CONFIG` keeps working unchanged.
